Set up testing environment
If the SPACES_TMUX_SOCKET_NAME environment variable is set, spaces uses
that socket. This allows test code to have a separate socket, and thus
isolated server with sessions. Tests will not interfere with normal tmux
sessions now that there is the ability to mock this environmental
variable.

test-env.spec.ts is a test that, with the SPACES_TMUX_SOCKET_NAME
variable set, confirms that new-session launches a single tmux session
without any other sessions present in the new socket.

This should allow for confirmation that the testing environment itself
works.