Remove old hard-coded config logic
Prior to this point, the source included logic for launching specific
environments that I used. This included the paths in which these
projects were contained, as well as environmental variables that I do
not want published.

As of this version, only the logic that reads from YAML files remains,
and no personal information of mine remains in the source any longer.

Therefore, the state of the code after this commit is the code which
will likely be the first version (initial commit) of the repository when
I finally upload this to GitHub (or wherever else it may wind up).

While I wish I could include the full history of this project in the
repo, I do not think that there is information in the commits prior to
this point that is valuable enough to warrant spending time isolating
from the code that contained personal information.

Lot's of clean-up, filling out of tests, and continued development is
expected from this point onwards, so I anticipate that, before long,
there will be plenty of commits and history in this repository that
follows this commit.

Best regards to all who may read this,
Henry